pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ar-nids"
version = "0.1.0"
description = "Adversarial-resilient network intrusion detection system scaffold."
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "OpenAI Codex" }]
dependencies = [
  "fastapi>=0.115",
  "uvicorn>=0.30",
  "numpy>=1.26",
  "pandas>=2.2",
  "scikit-learn>=1.5",
  "joblib>=1.3",
  "pydantic>=2.8",
  "pyyaml>=6.0.2",
  "prometheus-client>=0.21",
  "mlflow>=2.16",
]

[project.optional-dependencies]
train = [
  "tensorflow>=2.16",
  "shap>=0.46",
]
dev = [
  "pytest>=8.3",
  "httpx>=0.27",
]

[project.scripts]
ar-nids-train = "ar_nids.cli:train_entrypoint"
ar-nids-serve = "ar_nids.cli:serve_entrypoint"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
